+++ Documents/changelog.md
@@ -11,3 +11,14 @@ TODO:
 1. Update the total depending on the equation being performed
 2. Clear the text box when the first number is added
 3. Make a equal button lol forgot about it
+
+# Calculate Commit: 
+1. Basic calculate function. It works for the most part. 
+2. Better Comments on main.js
+3. Append equation to previous calcs (This is in very early stage)
+4. Append name for the equation and also add a name input box above the normal text box
+
+TODO: 
+1. Style the prev equation boxes a bit better
+2. Style the main div for prev equations
+3. Utilise a tracking array for the previous equations